User updates

+++ app/controllers/api/v1/users_controller.rb
@@ -17,22 +17,33 @@ class Api::V1::UsersController < ApplicationController
 
 	def update
 		user = User.find(params[:id])
-		if user.update(user_params)
+		updated_attributes = user == current_user ? self_update_user_params : user_params
+		if  authorized_for_update(current_user, user, params[:user]) && user.update(updated_attributes)
 			render json: user, status: 200, location: [:api, user]
 		else
+			user.errors[:credentials] = "Insufficient priviledges" unless authorized_for_update(current_user, user, params[:user])
 			render json: { errors: user.errors }, status: 422
 		end
 	end
 
 	def destroy
 		user = User.find(params[:id])
-		user.destroy
-		head 204
+		if authorized_for_deletion(current_user, user)
+			user.destroy
+			head 204
+		else
+			head 403	
+		end
+		
 	end
 
 	private 
 
 		def user_params
 	      params.require(:user).permit(:email, :password, :password_confirmation, :credentials)
+	    end
+
+	    def self_update_user_params
+	    	params.require(:user).permit(:email, :password, :password_confirmation)
 	    end
 end
